code_gen: Added function symbol table, that is used to provide function pointer for external functions.
Replaced hardcoded if statements encoding function indices for each call
instruction with the `FunctionRefTable`.

+++ code_gen/src/code_gen/instr.h
@@ -200,7 +200,7 @@ struct Instr {
 		struct {
 			InstrInputs args;
 			InstrIndex io_state;
-			uint8_t function_index;
+			uint16_t function_index;
 		} call_internal;
 
 		struct {
